requirements.txt
flask>=2.3.0
requests>=2.31.0
numpy>=1.24.0
rich>=13.0.0
colorama>=0.4.6
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
